feat: shared/ui 분리
