+++ src/shared/ui/card/index.ts
@@ -1,0 +1,1 @@
+export { Card, CardHeader, CardTitle, CardContent } from "./Card"
