+++ src/shared/ui/dialog/index.ts
@@ -1,0 +1,1 @@
+export { Dialog, DialogContent, DialogHeader, DialogTitle } from "./Dialog"
