+++ src/shared/ui/textarea/Textarea.tsx
@@ -1,0 +1,16 @@
+import * as React from "react"
+import { forwardRef } from "react"
+
+export const Textarea = forwardRef<HTMLTextAreaElement, React.TextareaHTMLAttributes<HTMLTextAreaElement>>(
+  ({ className, ...props }, ref) => {
+    return (
+      <textarea
+        className={`flex min-h-[150px] w-full rounded-md border border-gray-300 bg-white px-3 py-2 text-sm ring-offset-background placeholder:text-gray-500 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-blue-500 focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className || ""}`}
+        ref={ref}
+        {...props}
+      />
+    )
+  },
+)
+
+Textarea.displayName = "Textarea"
