+++ src/shared/ui/textarea/index.ts
@@ -1,0 +1,1 @@
+export { Textarea } from "./Textarea"
